+++ include/GafferScene/ScenePlug.h
@@ -160,19 +160,13 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 
 			/// Standard constructors, for modifying context on the current thread.
 			PathScope( const Gaffer::Context *context );
-			[[deprecated("Use faster pointer version")]]
-			PathScope( const Gaffer::Context *context, const ScenePath &scenePath );
 			PathScope( const Gaffer::Context *context, const ScenePath *scenePath );
 
 			/// Specialised constructors used to transfer state to TBB tasks. See
 			/// ThreadState documentation for more details.
 			PathScope( const Gaffer::ThreadState &threadState );
-			[[deprecated("Use faster pointer version")]]
-			PathScope( const Gaffer::ThreadState &threadState, const ScenePath &scenePath );
 			PathScope( const Gaffer::ThreadState &threadState, const ScenePath *scenePath );
 
-			[[deprecated("Use faster pointer version")]]
-			void setPath( const ScenePath &scenePath );
 			void setPath( const ScenePath *scenePath );
 		};
 
@@ -185,19 +179,13 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 
 			/// Standard constructors, for modifying context on the current thread.
 			SetScope( const Gaffer::Context *context );
-			[[deprecated("Use faster pointer version")]]
-			SetScope( const Gaffer::Context *context, const IECore::InternedString &setName );
 			SetScope( const Gaffer::Context *context, const IECore::InternedString *setName );
 
 			/// Specialised constructors used to transfer state to TBB tasks. See
 			/// ThreadState documentation for more details.
 			SetScope( const Gaffer::ThreadState &threadState );
-			[[deprecated("Use faster pointer version")]]
-			SetScope( const Gaffer::ThreadState &threadState, const IECore::InternedString &setName );
 			SetScope( const Gaffer::ThreadState &threadState, const IECore::InternedString *setName );
 
-			[[deprecated("Use faster pointer version")]]
-			void setSetName( const IECore::InternedString &setName );
 			void setSetName( const IECore::InternedString *setName );
 		};